sort out folder rename
